starting fN
