+++ pyigm/data/DLA/dla_fits.json
@@ -1,0 +1,16 @@
+{
+    "CreationDate": "2017-Jun-03",
+    "User": "xavier",
+    "atan_lz": {
+        "A": 0.23636363446712494,
+        "B": 0.16818182170391083,
+        "C": 2.8686869144439697,
+        "form": "A + B*atan(z-C)"
+    },
+    "dpow_fN": {
+        "Nd": 21.545454025268555,
+        "a3": -2.021212100982666,
+        "a4": -6.0,
+        "form": "(N/Nd)**aa with aa=a3 if N<Nd else aa=a4"
+    }
+}
